add review

+++ server/model/profile.model.js
@@ -9,8 +9,8 @@ module.exports = (sequelize, Sequelize) => {
         last_name: Sequelize.TEXT + ' CHARSET utf8 COLLATE utf8_general_ci',
         description: Sequelize.TEXT + ' CHARSET utf8 COLLATE utf8_general_ci',
         address_detail: Sequelize.TEXT + ' CHARSET utf8 COLLATE utf8_general_ci',
-        address_longitude: Sequelize.INTEGER,
-        address_latitude: Sequelize.INTEGER
+        address_longitude: Sequelize.DOUBLE,
+        address_latitude: Sequelize.DOUBLE
     });
     
     return Profile;
